backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
PyJWT[crypto]==2.10.1
confluent-kafka==2.8.2
